requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.6
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
attrs==25.3.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.3
dataclasses-json==0.6.7
diskcache==5.6.3
fastapi==0.115.12
filelock==3.18.0
frozenlist==1.6.0
fsspec==2025.5.1
greenlet==3.2.2
h11==0.16.0
hf-xet==1.1.2
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.32.2
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.0.350
langchain-community==0.0.20
langchain-core==0.2.43
langchain-ollama==0.3.3
langchain-postgres==0.0.6
langchain-text-splitters==0.2.4
langserve==0.0.6
langsmith==0.1.147
llama_cpp_python==0.3.9
MarkupSafe==3.0.2
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.4.4
mypy_extensions==1.1.0
networkx==3.5
numpy==1.26.4
ollama==0.5.1
orjson==3.10.18
packaging==23.2
pdfminer.six==20250506
pgvector==0.2.5
pillow==11.2.1
propcache==0.3.1
psycopg==3.2.9
psycopg-pool==3.2.6
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==1.10.22
pydantic-settings==2.9.1
pydantic_core==2.33.2
PyMuPDF==1.26.0
pypdf==5.5.0
pypdfium2==4.30.1
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sentence-transformers==4.1.0
setuptools==80.9.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sse-starlette==2.3.6
starlette==0.46.2
sympy==1.14.0
tenacity==8.5.0
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.52.3
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
yarl==1.20.0
zstandard==0.23.0
